requirements.txt
fastapi
uvicorn
PyGithub
requests
python-dotenv
pytest